pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-study-buddy"
version = "0.1.0"
description = "AI-powered study buddy application for generating and managing quizzes"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "jasonli", email = "jasonli@example.com" },
]
keywords = ["ai", "study", "quiz", "education", "streamlit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "groq>=0.4.0",
    "pandas>=1.5.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
requires-python = ">=3.12"

[project.urls]
Homepage = "https://github.com/yourusername/ai-study-buddy"
Repository = "https://github.com/yourusername/ai-study-buddy"
Documentation = "https://github.com/yourusername/ai-study-buddy#readme"
Issues = "https://github.com/yourusername/ai-study-buddy/issues"

[project.scripts]
ai-study-buddy = "application:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
